add unsafe pointer truncation check for stringslice
